requirements.txt
google-cloud-datastore==1.3.0
python-json-logger==0.1.8
pyyaml==3.12
